[Step07-Ex] feat: Add LiveData<Boolean> for the button enabled states
For the start/stop buttons: make them depend on the null/not-null state
of the tonight LiveData, i.e. whether or not we're currently tracking
a night's sleep.
For the clear button, make it depend on the isNotEmpty() state of the
nights LiveData (whether or not we do have records in the sleep table)